Fix clean up logic for QEMU VMs.
Previously, garbage resources of QEMU processes was not cleaned
properly.  This commit makes sure to clean up them.
